v0 Created file upload handler
At this commit only upload of a single file <32mb in size is supported. The general idea is to accept request, register a uuid that wont be used in parallel upload, stream the file to tmp directory, register file properly in files table and then move it to files root. All metadata such as folder, type, original name, owner is stored in db only.